Skip resetting clean environments already at the correct commit
Skips resetting an environment so long as it's already at the same
commit as upstream, and is not dirty (contains no uncommitted changes).
Does not check status of third party modules when determining whether
to update the environment or not, as this check is expensive.

Can be forced if needed, which will ensure third party modules are
also correct.