Improve concurrency and race conditions
* Lock the master repo while changes are being fetched
* Lock environments while they're being added/updated/removed
* Deploy updated environment code into a new directory, and manage
  a symlink to point at live copy of the code

These changes make the env manager more robust to concurrent access
if multiple changes are triggered in quick succession.

By deploying to a new directory and updating a symlink we get closer
to atomic environment updates, which means the puppet server should
not ever be trying to read from an environment in which the librarian
is busy updating modules. Under the previous approach, the master would
occasionally fail to compile a catalog, or serve garbage during the
update process.