Protect against race condition during stale clone cleanup
If two processes run concurrently, one might be constructing a new clone
directory, and the other might incorrectly detect that as a stale
environment and delete it. To mitigate the race condition, check whether
environment is locked at the point of considering it for deletion.

There's still a small window for a race here, but it's much reduced.